% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_boxscores}
\alias{historical_boxscores}
\title{historical_boxscores}
\format{
\subsection{\code{historical_boxscores}}{

A data frame 13212 rows and 33 columns
}
}
\source{
Generated by calling \code{generate_box_scores()} on all historical games
}
\usage{
historical_boxscores
}
\description{
Boxscore and plus/minus data for previous TPL seasons
}
\keyword{datasets}
