% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_possessions}
\alias{historical_possessions}
\title{historical_possessions}
\format{
\subsection{\code{historical_possessions}}{

A data frame 166304 rows and 13 columns
}
}
\source{
Generated by calling \code{create_detailed_possessions()} on all historical games
}
\usage{
historical_possessions
}
\description{
Detailed possession level data for previous TPL seasons
}
\keyword{datasets}
