% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s13_games}
\alias{s13_games}
\title{s13_games}
\format{
\subsection{\code{s13_games}}{

A data frame 75 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/574}
}
\usage{
s13_games
}
\description{
TPL season 13 game summary data
}
\keyword{datasets}
