% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s14_games}
\alias{s14_games}
\title{s14_games}
\format{
\subsection{\code{s14_games}}{

A data frame 84 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/596}
}
\usage{
s14_games
}
\description{
TPL season 14 game summary data
}
\keyword{datasets}
