% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s15_games}
\alias{s15_games}
\title{s15_games}
\format{
\subsection{\code{s15_games}}{

A data frame 77 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/629}
}
\usage{
s15_games
}
\description{
TPL season 15 game summary data
}
\keyword{datasets}
