% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s16_games}
\alias{s16_games}
\title{s16_games}
\format{
\subsection{\code{s13_games}}{

A data frame 30 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/652}
}
\usage{
s16_games
}
\description{
TPL season 16 game summary data
}
\keyword{datasets}
