% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s17_games}
\alias{s17_games}
\title{s17_games}
\format{
\subsection{\code{s13_games}}{

A data frame 36 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/694}
}
\usage{
s17_games
}
\description{
TPL season 17 game summary data
}
\keyword{datasets}
