% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s18_games}
\alias{s18_games}
\title{s18_games}
\format{
\subsection{\code{s18_games}}{

A data frame 52 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/703}
}
\usage{
s18_games
}
\description{
TPL season 18 game summary data
}
\keyword{datasets}
