% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s19_games}
\alias{s19_games}
\title{s19_games}
\format{
\subsection{\code{s19_games}}{

A data frame 65 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/728}
}
\usage{
s19_games
}
\description{
TPL season 19 game summary data
}
\keyword{datasets}
