% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s19_players}
\alias{s19_players}
\title{s19_players}
\format{
\subsection{\code{s19_games}}{

A data frame 134 rows and 5 columns
}
}
\source{
Generated manually from S19 draft spreadsheet
}
\usage{
s19_players
}
\description{
A list of all season 19 players with general info
}
\keyword{datasets}
