% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s20_games}
\alias{s20_games}
\title{s20_games}
\format{
\subsection{\code{s20_games}}{

A data frame 70 rows and 14 columns
}
}
\source{
\url{https://tplapp.onrender.com/games/739}
}
\usage{
s20_games
}
\description{
TPL season 20 game summary data
}
\keyword{datasets}
