% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s20_players}
\alias{s20_players}
\title{s20_players}
\format{
\subsection{\code{s20_games}}{

A data frame 114 rows and 5 columns
}
}
\source{
Generated manually from S20 draft spreadsheet
}
\usage{
s20_players
}
\description{
A list of all season 20 players with general info
}
\keyword{datasets}
