% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s21_boxscores}
\alias{s21_boxscores}
\title{s21_boxscores}
\format{
\subsection{\code{s21_boxscores}}{

A data frame 1389 rows and 14 columns
}
}
\source{
Generated by Viv using \url{https://tplapp.onrender.com/games/764}
}
\usage{
s21_boxscores
}
\description{
TPL season 21 boxscore data
}
\keyword{datasets}
