% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s21_game_IDs}
\alias{s21_game_IDs}
\title{s21_game_IDs}
\format{
\subsection{\code{s21_game_IDs}}{

A data frame 50 rows and 3 columns
}
}
\source{
Generated manually by Viv
}
\usage{
s21_game_IDs
}
\description{
A list of s21 games with game and team IDs
}
\keyword{datasets}
