% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s21_players}
\alias{s21_players}
\title{s21_players}
\format{
\subsection{\code{s21_games}}{

A data frame 126 rows and 1 columns
}
}
\source{
Generated manually from S21 draft spreadsheet
}
\usage{
s21_players
}
\description{
A list of all season 20 players with general info
}
\keyword{datasets}
